Added indicators for missed events (issue #2).
When an events time is less than the current time, the human time diff gets the "ago" indicator. Additionally, if doing cron is "no" and the current time is greater than an event's time, the event is printed in red. This should allow for very easily seeing if a number of events have been missed.