Add counts for core/custom events to top of page & move schedules up.
Rationale for moving schedules up:
If there are a lot of events, it would previously get a bit lost in between the custom and core events. As it's normally a short list, having it at the top seems more logical.

Also note that the event counts will count events with the same time & hook, but different arguments as individual events.